rename folder and fix imports
