--- omnia_timeseries/__init__.py
@@ -1,1 +1,0 @@
-from omnia_timeseries.api import TimeseriesAPI, TimeseriesEnvironment
