+++ src/omnia_timeseries_api/__init__.py
@@ -1,0 +1,1 @@
+from omnia_timeseries_api.api import TimeseriesAPI, TimeseriesEnvironment
