+++ src/omnia_timeseries_api/models.py
