+++ tests/test_api.py
@@ -4,8 +4,8 @@ import pytest
 from azure.identity._internal.msal_credentials import MsalCredential
 from azure.core.credentials import AccessToken
 from requests.models import Response
-from omnia_timeseries import TimeseriesAPI, TimeseriesEnvironment
-from omnia_timeseries.http_client import TimeseriesRequestFailedException
+from omnia_timeseries_api import TimeseriesAPI, TimeseriesEnvironment
+from omnia_timeseries_api.http_client import TimeseriesRequestFailedException
 
 
 class DummyCredentials(MsalCredential):
